Moved GET promotions logic into custom repository PromotionRepository
