+++ backend/ormconfig.json
@@ -1,6 +1,6 @@
 {
   "type": "postgres",
-  "host": "db",
+  "host": "localhost",
   "port": 5432,
   "username": "postgres",
   "password": "postgres",
