+++ backend/src/main/controller/PromotionController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import { PromotionRepository } from '../repository/PromotionRepository';
-import { getCustomRepository, SelectQueryBuilder } from 'typeorm';
+import { getCustomRepository } from 'typeorm';
 import { Promotion } from '../entity/Promotion';
 import { UserRepository } from '../repository/UserRepository';
 import { Discount } from '../entity/Discount';
@@ -8,10 +8,6 @@ import {
   PromotionDTO,
   PromotionValidation,
 } from '../validation/PromotionValidation';
-import {
-  PromotionQueryDTO,
-  PromotionQueryValidation,
-} from '../validation/PromotionQueryValidation';
 import { IdValidation } from '../validation/IdValidation';
 
 /* eslint-disable  @typescript-eslint/no-explicit-any */
@@ -28,77 +24,14 @@ export class PromotionController {
     next: NextFunction
   ): Promise<any> => {
     try {
-      let queryBuilder = getCustomRepository(PromotionRepository)
-        .createQueryBuilder('promotion')
-        .innerJoinAndSelect('promotion.discount', 'discount');
-
-      if (request.query) {
-        const promotionQuery: PromotionQueryDTO = await PromotionQueryValidation.schema.validateAsync(
-          request.query,
-          { abortEarly: false }
-        );
-        queryBuilder = this.applyQueryOptions(queryBuilder, promotionQuery);
-      }
-
-      const promotions = await queryBuilder.getMany();
+      const promotions = await getCustomRepository(
+        PromotionRepository
+      ).getAllPromotions(request.query);
       return response.send(promotions);
     } catch (e) {
       return next(e);
     }
   };
-
-  /**
-   * Depending on which properties are defined inside promotionQuery, we add those properties into our query for the queryBuilder to execute.
-   */
-  private applyQueryOptions(
-    queryBuilder: SelectQueryBuilder<Promotion>,
-    promotionQuery: PromotionQueryDTO
-  ): SelectQueryBuilder<Promotion> {
-    if (promotionQuery?.category) {
-      queryBuilder = queryBuilder.andWhere('promotion.category = :category', {
-        category: promotionQuery.category,
-      });
-    }
-
-    if (promotionQuery?.cuisine) {
-      queryBuilder = queryBuilder.andWhere('promotion.cuisine = :cuisine', {
-        cuisine: promotionQuery.cuisine,
-      });
-    }
-
-    if (promotionQuery?.discountType) {
-      queryBuilder = queryBuilder.andWhere('discount.type = :type', {
-        type: promotionQuery.discountType,
-      });
-    }
-
-    if (promotionQuery?.expirationDate) {
-      queryBuilder = queryBuilder.andWhere(
-        'promotion.expirationDate >= :date',
-        {
-          date: new Date(promotionQuery.expirationDate),
-        }
-      );
-    }
-
-    if (promotionQuery?.name) {
-      let name = promotionQuery.name;
-
-      // first escape % and _ (request needs to show it as https://stackoverflow.com/questions/17342671/pass-a-percent-sign-in-a-url-and-get-exact-value-of-it-using-php)
-      // this way users can query promotions that contain % or _
-      name = name.replace('%', '\\%');
-      name = name.replace('_', '\\_');
-
-      // add wildcards
-      name = `%${name.replace(/ /g, '%')}%`;
-
-      queryBuilder = queryBuilder.andWhere('promotion.name ilike :name', {
-        name: name,
-      });
-    }
-
-    return queryBuilder;
-  }
 
   /**
    * Retrieves a single promotion and its discount
